+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Yii2 multiple input change log
 1.2.4 in development
 --------------------
 
+- Bug #39: TabularInput: now new row does't copy values from  the most recent row
+
 1.2.3
 -----